+++ simple_manga_downloader/modules/__init__.py
@@ -1,0 +1,6 @@
+from .mangadex_org import Mangadex
+from .mangaseeonline_us import Mangasee
+from .mangatown_com import Mangatown
+from .heavenmanga_org import Heavenmanga
+from .mangakakalot_com import Mangakakalot
+from .config_parser import Config