+++ simple_manga_downloader/modules/mangatown_com.py
@@ -1,28 +1,34 @@
 import requests
 from bs4 import BeautifulSoup
+from ..decorators import request_exception_handler
 import re
 
 
 class Mangatown():
     def __init__(self, link, directory):
-        self.scraper = None
+        self.session = requests.Session()
         self.site = "mangatown.com"
         self.folder = directory
         self.manga_link = link
         self.base_link = "https://www.mangatown.com/"
+        self.cover_url = None
 
-    def get_chapters(self, title_return):
+    @request_exception_handler
+    def get_chapters(self, title_return=False):
         '''Gets the list of available chapters
         title_return=True will not create the chapters dict,
         used if only title is needed'''
 
-        r = requests.get(self.manga_link, timeout=5)
+        r = self.session.get(self.manga_link, timeout=5)
         r.raise_for_status()
         soup = BeautifulSoup(r.text, "html.parser")
 
         self.series_title = soup.find(class_="title-top").string
         if title_return:
             return True
+        thumb = soup.find(class_="detail_info")
+        if thumb:
+            self.cover_url = thumb.img["src"]
         self.manga_dir = self.folder / self.series_title
 
         chapters_soup = soup.find(class_="chapter_list").find_all("li")
@@ -50,10 +56,11 @@ class Mangatown():
                                      "title": chapter_title}
         return True
 
+    @request_exception_handler
     def get_info(self, ch):
         '''Gets the needed data abut the chapters from the site'''
 
-        r = requests.get(self.chapters[ch]["link"], timeout=5)
+        r = self.session.get(self.chapters[ch]["link"], timeout=5)
         r.raise_for_status()
         soup = BeautifulSoup(r.text, "html.parser")
 
@@ -62,7 +69,7 @@ class Mangatown():
 
         image_links = []
         for page in pages:
-            page_r = requests.get(page, timeout=5)
+            page_r = self.session.get(page, timeout=5)
             page_r.raise_for_status()
             page_soup = BeautifulSoup(page_r.text, "html.parser")
             img_link = page_soup.find(id="image")["src"]