+++ template/module_template.py
@@ -1,29 +1,34 @@
 # Cloud flare
-import cfscrape
-import requests.exceptions
+# import cfscrape
+# import requests.exceptions
 
 # No cloudflare
 import requests
 
 from bs4 import BeautifulSoup
+from ..decorators import request_exception_handler
 
 
 class ClassName:
     def __init__(self, link, directory):
-        self.scraper = None
+        self.session = requests.Session()
         self.site = "site.com"
         self.folder = directory
         self.manga_link = link
         self.base_link = "https://site.com"
+        self.cover_url = None
 
-    def get_chapters(self, title_return):
+    @request_exception_handler
+    def get_chapters(self, title_return=False):
         '''Gets the list of available chapters
         title_return=True will not create the chapters dict,
         used if only title is needed'''
 
-        r = requests.get(self.manga_link, timeout=5)
+        r = self.session.get(self.manga_link, timeout=5)
         r.raise_for_status()
         soup = BeautifulSoup(r.text, "html.parser")
+
+        self.cover_url = "get cover here"
 
         self.series_title = "find title"
         self.manga_dir = self.folder / self.series_title
@@ -46,11 +51,12 @@ class ClassName:
                                   "title": None}
         return True
 
+    @request_exception_handler
     def get_info(self, ch):
         '''Gets the needed data abut the chapters from the site'''
         link = self.chapters[ch]["link"]
 
-        r = requests.get(link, timeout=5)
+        r = self.session.get(link, timeout=5)
         r.raise_for_status()
         soup = BeautifulSoup(r.text, "html.parser")